dom/base/test/test_appname_override.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=939445
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 939445 - general.appname.override</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=939445">Mozilla Bug 939445</a>
  <script type="application/javascript">

  function runTest() {
    is(navigator.appName, "hello", "general.appname.override not working");

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [["general.appname.override", "hello"]]}, runTest);
  </script>
</body>
</html>
